requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
protobuf>=4.25.0
Pillow>=10.0.0 